Finished Text Box
Added functionality for a text box that could be resized and have its contents changed. Also added a button to create text boxes.  Adjusted the pen and eraser buttons so that they wouldn't interfere with moving a text box, and had to create a temporary eraser function to do this.